Refactor: Remove authentication routes from the app.py file and create auth_routes.py
